feature: implement the innards of the logger middleware

+++ logly/Models/Response.cs
@@ -3,7 +3,7 @@
     internal class Response
     {
         public int StatusScode { get; set; }
-        public int ResponseLength { get; set; }
+        public long ResponseLength { get; set; }
         public double ResponseTime { get; set; }
     }
 }
